A checked way of accessing user memory
Basically, instead of maintaining a map of virtual memory that is valid and checking that first so that a page fault never happens, we just "catch" page faults and gracefully return an error. We already have page tables, so why maintain another map? We can let the CPU do the checking for us. This improves performance, reduces the complexity of code that accesses user memory, and reduces memory usage.

Also enables SMEP if supported. Enables SMAP if supported and allows accessing user mem only when needed.

The log syscall handler was improved. It now uses the checked user mem access function. It also buffers the string, literally copying from user mem to kernel mem, which fixes the issue of another thread modifying the &str to be invalid utf8 after the utf8 check happened.